Moved vault ciphertext encryption module to vault dir.
