+++ lib/control/v1/vault.js
@@ -1,4 +1,4 @@
-const encryptCipher = require('../../encrypt_ciphertext');
+const encryptCipher = require('../../vault/ciphertext');
 
 module.exports = function Encrypt() {
   return (req, res, next) => {
