+++ lib/vault/ciphertext.js
